Внедрил liquebase. Создал в resources scripts и changelog-master.yml. Распеределил по пакетам все классы сервиса и model для удобства
