+++ src/main/java/skypro/teamwork/telegram_bot_for_shelter/repository/CatRepository.java
@@ -1,7 +1,7 @@
 package skypro.teamwork.telegram_bot_for_shelter.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import skypro.teamwork.telegram_bot_for_shelter.model.Cat;
+import skypro.teamwork.telegram_bot_for_shelter.model.cat.Cat;
 
 public interface CatRepository extends JpaRepository<Cat, Long> {
 }
