+++ src/main/java/skypro/teamwork/telegram_bot_for_shelter/repository/ReportCatRepository.java
@@ -1,7 +1,7 @@
 package skypro.teamwork.telegram_bot_for_shelter.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import skypro.teamwork.telegram_bot_for_shelter.model.ReportCat;
+import skypro.teamwork.telegram_bot_for_shelter.model.cat.ReportCat;
 
 public interface ReportCatRepository extends JpaRepository<ReportCat, Long> {
 }
