+++ src/main/java/skypro/teamwork/telegram_bot_for_shelter/service/cats/CatService.java
@@ -1,7 +1,7 @@
-package skypro.teamwork.telegram_bot_for_shelter.service;
+package skypro.teamwork.telegram_bot_for_shelter.service.cats;
 
 import org.springframework.stereotype.Service;
-import skypro.teamwork.telegram_bot_for_shelter.model.Cat;
+import skypro.teamwork.telegram_bot_for_shelter.model.cat.Cat;
 import skypro.teamwork.telegram_bot_for_shelter.repository.CatRepository;
 
 import java.util.Collection;
