+++ src/main/resources/application.properties
@@ -1,7 +1,12 @@
-spring.datasource.url=
-spring.datasource.username=
-spring.datasource.password=
-bot.name=
-bot.token=
+spring.datasource.url=jdbc:postgresql://localhost:5432/shelter
+spring.datasource.username=Azira
+spring.datasource.password=Azira
+spring.jpa.hibernate.ddl-auto=update
+
+bot.name=homework_for_sky_telegram_bot
+bot.token=6093682784:AAFVvmD87h03LtRCAEeRIwilxe3PIUxGYpE
+
 path.to.photocats.folder=photocats
 path.to.photocatsreport.folder=photocatsreport
+
+spring.liquibase.change-log=classpath:liquibase/changelog-master.yml
