+++ src/main/resources/liquibase/changelog-master.yml
@@ -1,0 +1,3 @@
+databaseChangeLog:
+  - include:
+      file: liquibase/scripts/cat.sql
