+++ src/main/resources/liquibase/scripts/cat.sql
@@ -1,0 +1,10 @@
+-- liquibase formatted sql
+-- changeset dsazonov: 1
+CREATE TABLE cat(
+);
+CREATE TABLE photo_cats(
+);
+CREATE TABLE photo_cats_report(
+);
+CREATE TABLE report_cat(
+);
